Some minor improvements attempting to standardise the results between the two implementations

+++ omp3/bright.h
@@ -11,8 +11,8 @@ void handle_particles(
     const int global_nx, const int global_ny, const int nx, const int ny, 
     const int x_off, const int y_off, const int initial, const double dt, 
     const int* neighbours, const double* density, const double* edgex, 
-    const double* edgey, const double* edgedx, const double* edgedy, int* facets, 
-    int* collisions, int* nparticles_sent, uint64_t* master_key, 
+    const double* edgey, const double* edgedx, const double* edgedy, uint64_t* facets, 
+    uint64_t* collisions, int* nparticles_sent, uint64_t* master_key, 
     const int ntotal_particles, const int nparticles_to_process, 
     int* nparticles, Particle* particles_start, CrossSection* cs_scatter_table, 
     CrossSection* cs_absorb_table, double* scalar_flux_tally, 
@@ -25,8 +25,8 @@ int handle_particle(
     const int initial, const int ntotal_particles, const double* density, 
     const double* edgex, const double* edgey, const double* edgedx, 
     const double* edgedy, const CrossSection* cs_scatter_table, 
-    const CrossSection* cs_absorb_table, int* nparticles_sent, int* facets, 
-    int* collisions, Particle* particle, 
+    const CrossSection* cs_absorb_table, int* nparticles_sent, uint64_t* facets, 
+    uint64_t* collisions, Particle* particle, 
     double* scalar_flux_tally, double* energy_deposition_tally, RNPool* rn_pool);
 
 // Calculate the distance to the next facet
